requirements.txt
cycler==0.11.0
h5py==3.8.0
keras==2.10.0
Keras-Preprocessing==1.1.2
keyboard==0.13.5
matplotlib==3.7.1
matplotlib-inline==0.1.6
menuinst==1.4.18
numpy==1.23.5
opencv-python==4.7.0.72
Pillow==9.5.0
scikit-learn==1.2.2
scipy==1.10.1
tensorflow==2.10.1
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
